Calculate summary statistics

Calculate epidemiological summary statistics such as case (e.g. injury) incidence and case burden (see Bahr et al. 2018), including total number of cases, number of days lost due to this event, total time of exposure etc., by means of a (widely used) Poisson method, negative binomial, zero-inflated poisson or zero-inflated negative binomial, on a athlete and overall basis.

Usage

calc_summary(
  injd,
  by = NULL,
  overall = TRUE,
  method = c("poisson", "negbin", "zinfpois", "zinfnb"),
  conf_level = 0.95,
  scale = TRUE,
  quiet = FALSE
)

Arguments

injd

injd S3 object (see prepare_all()).

by

Character specifying the name of the column according to which compute summary statistics. It should refer to a (categorical) variable that describes a grouping factor (e.g. "type of case or injury", "injury location", "sports club"). Optional, defaults to NULL.

overall

Logical, whether to calculate overall (for all the cohort) or athlete-wise summary statistic (i.e. number of cases per cohort of per athlete). Defaults to TRUE.

method

Method to estimate the incidence (burden) rate. One of "poisson", "negbin", "zinfpois" or "zinfnb"; that stand for Poisson method, negative binomial method, zero-inflated Poisson and zero-inflated negative binomial.

conf_level

Confidence level (defaults to 0.95).

scale

Logical, whether to transform the incidence and burden rates output according to the unit of exposure (defaults to TRUE).

quiet

Logical, whether or not to silence the warning messages (defaults to FALSE).

Value

A data frame comprising of overall or athlete-wise epidemiological summary statistics, that it's made up of the following columns:

totalexpo: total exposure that the athlete has been under risk of suffering a sports-related health problem.

ncases: number of sports-related health problems suffered by the athlete or overall in the team/cohort over the given period specified by the injd data frame.

ndayslost: number of days lost by the athlete or overall in the team/cohort due to the sports-related health problem over the given period specified by the injd data frame.

mean_dayslost: average of number of days lost (i.e. ndayslost) athlete-wise or overall in the team/cohort.

median_dayslost: median of number of days lost (i.e. ndayslost) athlete-wise or overall in the team/cohort.

qt25_dayslost and qt75_dayslost: interquartile range of number of days lost (i.e. ndayslost) athlete-wise or overall in the team/cohort.

incidence: case incidence rate, number of cases per unit of exposure.

burden: case burden rate, number of days lost per unit of exposure.

incidence_sd and burden_sd: estimated standard deviation, by the specified method argument, of case incidence (incidence) and case burden (burden).

incidence_lower and burden_lower: lower bound of, for example, 95% confidence interval (if conf_level = 0.95) of case incidence (incidence) and case burden (burden).

incidence_upper and burden_upper: the same (as above item) applies but for the upper bound.

Apart from this column names, they may further include these other columns depending on the user's specifications to the function:

by: only if it is specified as an argument to function.

percent_ncases: percentage (%) of number of cases of that type relative to all types of cases (if by specified).

percent_dayslost: percentage (%) of number of days lost because of cases of that type relative to the total number of days lost because of all types of cases (if by specified).

References

Bahr R., Clarsen B., & Ekstrand J. (2018). Why we should focus on the burden of injuries and illnesses, not just their incidence. British Journal of Sports Medicine, 52(16), 1018–1021. doi:10.1136/bjsports-2017-098160

Waldén M., Mountjoy M., McCall A., Serner A., Massey A., Tol J. L., ... & Andersen T. E. (2023). Football-specific extension of the IOC consensus statement: methods for recording and reporting of epidemiological data on injury and illness in sport 2020. British journal of sports medicine.

Examples

calc_summary(injd)
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> # A tibble: 1 × 15
#>   totalexpo ncases incidence incidence_sd incidence_lower incidence_upper
#>       <dbl>  <dbl>     <dbl>        <dbl>           <dbl>           <dbl>
#> 1     74690     82      9.88         1.09            7.74            12.0
#> # ℹ 9 more variables: ndayslost <dbl>, mean_dayslost <dbl>,
#> #   median_dayslost <dbl>, qt25_dayslost <dbl>, qt75_dayslost <dbl>,
#> #   burden <dbl>, burden_sd <dbl>, burden_lower <dbl>, burden_upper <dbl>
calc_summary(injd, overall = FALSE)
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> # A tibble: 28 × 16
#>    person_id             totalexpo ncases incidence incidence_sd incidence_lower
#>    <fct>                     <dbl>  <dbl>     <dbl>        <dbl>           <dbl>
#>  1 adam-lallana                700      6     77.1         31.5            15.4 
#>  2 alberto-moreno             1264      1      7.12         7.12           -6.84
#>  3 alex-oxlade-chamberl…      1483      1      6.07         6.07           -5.83
#>  4 alisson                    3420      0      0            0               0   
#>  5 andrew-robertson           5162      5      8.72         3.90            1.08
#>  6 daniel-sturridge            927      3     29.1         16.8            -3.83
#>  7 danny-ings                  265      0      0            0               0   
#>  8 dejan-lovren               3109      6     17.4          7.09            3.47
#>  9 divock-origi                366      1     24.6         24.6           -23.6 
#> 10 dominic-solanke             581      0      0            0               0   
#> # ℹ 18 more rows
#> # ℹ 10 more variables: incidence_upper <dbl>, ndayslost <dbl>,
#> #   mean_dayslost <dbl>, median_dayslost <dbl>, qt25_dayslost <dbl>,
#> #   qt75_dayslost <dbl>, burden <dbl>, burden_sd <dbl>, burden_lower <dbl>,
#> #   burden_upper <dbl>
calc_summary(injd, by = "injury_type")
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> # A tibble: 5 × 18
#>   injury_type ncases percent_ncases totalexpo incidence incidence_sd
#>   <chr>        <dbl>          <dbl>     <dbl>     <dbl>        <dbl>
#> 1 Bone            11           13.4     74690      1.33        0.400
#> 2 Concussion      16           19.5     74690      1.93        0.482
#> 3 Ligament         9           11.0     74690      1.08        0.361
#> 4 Muscle          25           30.5     74690      3.01        0.602
#> 5 Unknown         21           25.6     74690      2.53        0.552
#> # ℹ 12 more variables: incidence_lower <dbl>, incidence_upper <dbl>,
#> #   ndayslost <dbl>, percent_ndayslost <dbl>, mean_dayslost <dbl>,
#> #   median_dayslost <dbl>, qt25_dayslost <dbl>, qt75_dayslost <dbl>,
#> #   burden <dbl>, burden_sd <dbl>, burden_lower <dbl>, burden_upper <dbl>
calc_summary(injd, by = "injury_type", overall = FALSE)
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#> Warning: 
#>  Exposure time unit is matches_minutes
#>  Case incidence and case burden are calculated per 100 athlete-matches of exposure (i.e. 90 minutes times 100)
#> 
#>                  person_id injury_type totalexpo ncases incidence incidence_sd
#> 1             adam-lallana        Bone       700      0  0.000000     0.000000
#> 2             adam-lallana  Concussion       700      3 38.571429    22.269225
#> 3             adam-lallana    Ligament       700      0  0.000000     0.000000
#> 4             adam-lallana      Muscle       700      3 38.571429    22.269225
#> 5             adam-lallana     Unknown       700      0  0.000000     0.000000
#> 6           alberto-moreno        Bone      1264      0  0.000000     0.000000
#> 7           alberto-moreno  Concussion      1264      0  0.000000     0.000000
#> 8           alberto-moreno    Ligament      1264      1  7.120253     7.120253
#> 9           alberto-moreno      Muscle      1264      0  0.000000     0.000000
#> 10          alberto-moreno     Unknown      1264      0  0.000000     0.000000
#> 11 alex-oxlade-chamberlain        Bone      1483      0  0.000000     0.000000
#> 12 alex-oxlade-chamberlain  Concussion      1483      0  0.000000     0.000000
#> 13 alex-oxlade-chamberlain    Ligament      1483      1  6.068780     6.068780
#> 14 alex-oxlade-chamberlain      Muscle      1483      0  0.000000     0.000000
#> 15 alex-oxlade-chamberlain     Unknown      1483      0  0.000000     0.000000
#> 16                 alisson        Bone      3420      0  0.000000     0.000000
#> 17                 alisson  Concussion      3420      0  0.000000     0.000000
#> 18                 alisson    Ligament      3420      0  0.000000     0.000000
#> 19                 alisson      Muscle      3420      0  0.000000     0.000000
#> 20                 alisson     Unknown      3420      0  0.000000     0.000000
#> 21        andrew-robertson        Bone      5162      1  1.743510     1.743510
#> 22        andrew-robertson  Concussion      5162      2  3.487021     2.465696
#> 23        andrew-robertson    Ligament      5162      0  0.000000     0.000000
#> 24        andrew-robertson      Muscle      5162      1  1.743510     1.743510
#> 25        andrew-robertson     Unknown      5162      1  1.743510     1.743510
#> 26        daniel-sturridge        Bone       927      0  0.000000     0.000000
#> 27        daniel-sturridge  Concussion       927      0  0.000000     0.000000
#> 28        daniel-sturridge    Ligament       927      0  0.000000     0.000000
#> 29        daniel-sturridge      Muscle       927      1  9.708738     9.708738
#> 30        daniel-sturridge     Unknown       927      2 19.417476    13.730229
#> 31              danny-ings        Bone       265      0  0.000000     0.000000
#> 32              danny-ings  Concussion       265      0  0.000000     0.000000
#> 33              danny-ings    Ligament       265      0  0.000000     0.000000
#> 34              danny-ings      Muscle       265      0  0.000000     0.000000
#> 35              danny-ings     Unknown       265      0  0.000000     0.000000
#> 36            dejan-lovren        Bone      3109      1  2.894821     2.894821
#> 37            dejan-lovren  Concussion      3109      1  2.894821     2.894821
#> 38            dejan-lovren    Ligament      3109      0  0.000000     0.000000
#> 39            dejan-lovren      Muscle      3109      3  8.684464     5.013978
#> 40            dejan-lovren     Unknown      3109      1  2.894821     2.894821
#> 41            divock-origi        Bone       366      0  0.000000     0.000000
#> 42            divock-origi  Concussion       366      0  0.000000     0.000000
#> 43            divock-origi    Ligament       366      0  0.000000     0.000000
#> 44            divock-origi      Muscle       366      0  0.000000     0.000000
#> 45            divock-origi     Unknown       366      1 24.590164    24.590164
#> 46         dominic-solanke        Bone       581      0  0.000000     0.000000
#> 47         dominic-solanke  Concussion       581      0  0.000000     0.000000
#> 48         dominic-solanke    Ligament       581      0  0.000000     0.000000
#> 49         dominic-solanke      Muscle       581      0  0.000000     0.000000
#> 50         dominic-solanke     Unknown       581      0  0.000000     0.000000
#> 51                emre-can        Bone      2094      0  0.000000     0.000000
#> 52                emre-can  Concussion      2094      0  0.000000     0.000000
#> 53                emre-can    Ligament      2094      0  0.000000     0.000000
#> 54                emre-can      Muscle      2094      0  0.000000     0.000000
#> 55                emre-can     Unknown      2094      1  4.297994     4.297994
#> 56                 fabinho        Bone      2013      1  4.470939     4.470939
#> 57                 fabinho  Concussion      2013      0  0.000000     0.000000
#> 58                 fabinho    Ligament      2013      0  0.000000     0.000000
#>    incidence_lower incidence_upper ndayslost mean_dayslost median_dayslost
#> 1         0.000000        0.000000         0       0.00000               0
#> 2        -5.075450       82.218307        97      32.33333              25
#> 3         0.000000        0.000000         0       0.00000               0
#> 4        -5.075450       82.218307       205      68.33333              45
#> 5         0.000000        0.000000         0       0.00000               0
#> 6         0.000000        0.000000         0       0.00000               0
#> 7         0.000000        0.000000         0       0.00000               0
#> 8        -6.835187       21.075693        50      50.00000              50
#> 9         0.000000        0.000000         0       0.00000               0
#> 10        0.000000        0.000000         0       0.00000               0
#> 11        0.000000        0.000000         0       0.00000               0
#> 12        0.000000        0.000000         0       0.00000               0
#> 13       -5.825810       17.963369       316     316.00000             316
#> 14        0.000000        0.000000         0       0.00000               0
#> 15        0.000000        0.000000         0       0.00000               0
#> 16        0.000000        0.000000         0       0.00000               0
#> 17        0.000000        0.000000         0       0.00000               0
#> 18        0.000000        0.000000         0       0.00000               0
#> 19        0.000000        0.000000         0       0.00000               0
#> 20        0.000000        0.000000         0       0.00000               0
#> 21       -1.673707        5.160728         3       3.00000               3
#> 22       -1.345655        8.319696         6       3.00000               3
#> 23        0.000000        0.000000         0       0.00000               0
#> 24       -1.673707        5.160728        12      12.00000              12
#> 25       -1.673707        5.160728         1       1.00000               1
#> 26        0.000000        0.000000         0       0.00000               0
#> 27        0.000000        0.000000         0       0.00000               0
#> 28        0.000000        0.000000         0       0.00000               0
#> 29       -9.320039       28.737514        16      16.00000              16
#> 30       -7.493278       46.328230       106      53.00000              53
#> 31        0.000000        0.000000         0       0.00000               0
#> 32        0.000000        0.000000         0       0.00000               0
#> 33        0.000000        0.000000         0       0.00000               0
#> 34        0.000000        0.000000         0       0.00000               0
#> 35        0.000000        0.000000         0       0.00000               0
#> 36       -2.778924        8.568567        72      72.00000              72
#> 37       -2.778924        8.568567         9       9.00000               9
#> 38        0.000000        0.000000         0       0.00000               0
#> 39       -1.142752       18.511681        70      23.33333              21
#> 40       -2.778924        8.568567         9       9.00000               9
#> 41        0.000000        0.000000         0       0.00000               0
#> 42        0.000000        0.000000         0       0.00000               0
#> 43        0.000000        0.000000         0       0.00000               0
#> 44        0.000000        0.000000         0       0.00000               0
#> 45      -23.605672       72.786000         5       5.00000               5
#> 46        0.000000        0.000000         0       0.00000               0
#> 47        0.000000        0.000000         0       0.00000               0
#> 48        0.000000        0.000000         0       0.00000               0
#> 49        0.000000        0.000000         0       0.00000               0
#> 50        0.000000        0.000000         0       0.00000               0
#> 51        0.000000        0.000000         0       0.00000               0
#> 52        0.000000        0.000000         0       0.00000               0
#> 53        0.000000        0.000000         0       0.00000               0
#> 54        0.000000        0.000000         0       0.00000               0
#> 55       -4.125920       12.721908        68      68.00000              68
#> 56       -4.291940       13.233818        11      11.00000              11
#> 57        0.000000        0.000000         0       0.00000               0
#> 58        0.000000        0.000000         0       0.00000               0
#>    qt25_dayslost qt75_dayslost      burden  burden_sd burden_lower burden_upper
#> 1            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 2           18.5          42.5 1247.142857 126.628172  998.9562011  1495.329513
#> 3            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 4           44.0          81.0 2635.714286 184.086271 2274.9118249  2996.516747
#> 5            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 6            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 7            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 8           50.0          50.0  356.012658  50.347793  257.3327973   454.692519
#> 9            0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 10           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 11           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 12           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 13         316.0         316.0 1917.734322 107.880984 1706.2914787  2129.177166
#> 14           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 15           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 16           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 17           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 18           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 19           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 20           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 21           3.0           3.0    5.230531   3.019848   -0.6882632    11.149325
#> 22           2.5           3.5   10.461062   4.270711    2.0906228    18.831500
#> 23           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 24          12.0          12.0   20.922123   6.039697    9.0845351    32.759711
#> 25           1.0           1.0    1.743510   1.743510   -1.6737071     5.160728
#> 26           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 27           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 28           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 29          16.0          16.0  155.339806  38.834951   79.2246996   231.454912
#> 30          52.0          54.0 1029.126214  99.957574  833.2129682  1225.039459
#> 31           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 32           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 33           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 34           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 35           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 36          72.0          72.0  208.427147  24.563375  160.2838170   256.570477
#> 37           9.0           9.0   26.053393   8.684464    9.0321558    43.074631
#> 38           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 39          17.0          28.5  202.637504  24.219814  155.1675405   250.107468
#> 40           9.0           9.0   26.053393   8.684464    9.0321558    43.074631
#> 41           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 42           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 43           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 44           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 45           5.0           5.0  122.950820  54.985278   15.1816548   230.719984
#> 46           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 47           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 48           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 49           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 50           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 51           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 52           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 53           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 54           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 55          68.0          68.0  292.263610  35.442169  222.7982361   361.728985
#> 56          11.0          11.0   49.180328  14.828427   20.1171454    78.243510
#> 57           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#> 58           0.0           0.0    0.000000   0.000000    0.0000000     0.000000
#>  [ reached 'max' / getOption("max.print") -- omitted 82 rows ]