Injury summary statistics

Calculate injury summary statistics such as injury incidence and injury burden (see Bahr et al. 20), including total number of injuries, number of days lost due to injury, total time of exposure etc., by means of a classic-standard (Poisson) method, negative binomial, zero-inflated poisson or zero-inflated negative binomial, on a player and overall basis.

Usage

injsummary(
  injd,
  var_type_injury = NULL,
  method = c("poisson", "negbin", "zinfpois", "zinfnb"),
  conf_level = 0.95,
  quiet = FALSE
)

Arguments

injd

injd S3 object (see prepare_all()).

var_type_injury

Character specifying the name of the column according to which compute injury summary statistics. It should refer to a (categorical) variable that describes the "type of injury". Optional, defaults to NULL.

method

Method to estimate injury incidence and injury burden. One of "poisson", "negbin", "zinfpois" or "zinfnb"; characters that stand for Poisson method, negative binomial method, zero-inflated Poisson and zero-inflated negative binomial.

conf_level

Confidence level (default = 0.95).

quiet

Logical, whether or not to silence the warning messages (defaults to FALSE).

Value

A list of two data frames comprising player-wise and overall injury summary statistics, respectively, that constitute an injds

S3 object. Both of them made up of the following columns:

ninjuries: number of injuries sustained by the player or overall in the team over the given period specified by the injd data frame.

ndayslost: number of days lost by the player or overall in the team due to injury over the given period specified by the injd data frame.

median_dayslost: median of number of days lost (i.e. ndayslost) playerwise or overall in the team.

iqr_dayslost: interquartile range of number of days lost (i.e. ndayslost) playerwise or overall in the team.

totalexpo: total exposure that the player has been under risk of sustaining an injury.

injincidence: injury incidence, number of injuries per unit of exposure.

injburden: injury burden, number of days lost per unit of exposure.

var_type_injury: only if it is specified as an argument to function.

Apart from this column names, they may further include these other columns depending on the user's specifications to the function:

percent_ninjuries: percentage (%) of number of injuries of that type relative to all types of injuries (if var_type_injury specified).

percent_dayslost: percentage (%) of number of days lost because of injuries of that type relative to the total number of days lost because of all types of injuries (if var_type_injury specified).

injincidence_sd and injburden_sd: estimated standard deviation, by the specified method parameter, of injury incidence (injincidence) and injury burden (injburden), for the overall injury summary statistics (the 2nd element of the function output).

injincidence_lower and injburden_lower: lower bound of, for example, 95% confidence interval (if conf_level = 0.95) of injury incidence (injincidence) and injury burden (injburden), for the overall injury summary statistics (the 2nd element of the function output).

injincidence_upper and injburden_upper: the same (as above item) applies but for the upper bound.

References

Bahr R., Clarsen B., & Ekstrand J. (2018). Why we should focus on the burden of injuries and illnesses, not just their incidence. British Journal of Sports Medicine, 52(16), 1018–1021. https://doi.org/10.1136/bjsports-2017-098160

Examples

# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

injsummary(injd)
#> Warning: 
#> * Exposure time unit is matches_minutes
#>   So... Injury incidence and injury burden are calculated per 100 player-matches of exposure (90 minutes times 100)
#> 
#> Warning: 
#> * Exposure time unit is matches_minutes
#>   So... Injury incidence and injury burden are calculated per 100 player-matches of exposure (90 minutes times 100)
#> 
#> $playerwise
#> # A tibble: 28 × 8
#>    player                ninju…¹ ndays…² media…³ iqr_d…⁴ total…⁵ injin…⁶ injbu…⁷
#>    <fct>                   <dbl>   <dbl>   <dbl> <chr>     <dbl>   <dbl>   <dbl>
#>  1 adam-lallana                6     302    43   18.5-5…     700   77.1   3883. 
#>  2 alberto-moreno              1      50    25   12.5-3…    1264    7.12   356. 
#>  3 alex-oxlade-chamberl…       1     316   316   316-316    1483    6.07  1918. 
#>  4 alisson                     0       0     0   0-0        3420    0        0  
#>  5 andrew-robertson            5      22     2.5 1.25-3…    5162    8.72    38.4
#>  6 daniel-sturridge            3     122    33.5 12-52       927   29.1   1184. 
#>  7 danny-ings                  0       0     0   0-0         265    0        0  
#>  8 dejan-lovren                6     160    13   9-28.5     3109   17.4    463. 
#>  9 divock-origi                1       5     2.5 1.25-3…     366   24.6    123. 
#> 10 dominic-solanke             0       0     0   0-0         581    0        0  
#> # … with 18 more rows, and abbreviated variable names ¹​ninjuries, ²​ndayslost,
#> #   ³​median_dayslost, ⁴​iqr_dayslost, ⁵​totalexpo, ⁶​injincidence, ⁷​injburden
#> 
#> $overall
#> # A tibble: 1 × 13
#>   ninjuries ndayslost median_d…¹ iqr_d…² total…³ injin…⁴ injbu…⁵ injin…⁶ injbu…⁷
#>       <dbl>     <dbl>      <dbl> <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1        82      2049        7.5 1-20.25   74690    9.88    247.    1.09    5.45
#> # … with 4 more variables: injincidence_lower <dbl>, injincidence_upper <dbl>,
#> #   injburden_lower <dbl>, injburden_upper <dbl>, and abbreviated variable
#> #   names ¹​median_dayslost, ²​iqr_dayslost, ³​totalexpo, ⁴​injincidence,
#> #   ⁵​injburden, ⁶​injincidence_sd, ⁷​injburden_sd
#> 
#> attr(,"class")
#> [1] "injds" "list" 
#> attr(,"unit_exposure")
#> [1] "matches_minutes"
#> attr(,"unit_timerisk")
#> [1] "100 player-match"