injurytools package

Sports Injury Data analysis aims to identify and describe the magnitude of the injury problem, and to gain more insights (e.g. determine potential risk factors) by statistical modelling approaches. The 'injurytools' package provides standardized routines and utilities that simplify such analyses. It offers functions for data preparation, informative visualizations and descriptive and model-based analyses.

See also

Author

Maintainer: Lore Zumeta Olaskoaga lorezumeta@gmail.com (ORCID)

Other contributors:

Dae-Jin Lee daelee@faculty.ie.edu (ORCID) [contributor]