Injury summary statistics

Calculate injury summary statistics such as injury incidence and injury burden (REF-ak), including total number of injuries, number of days lost due to injury, total time of exposure etc., by means of a classic-standard (Poisson) method, negative binomial, zero-inflated poisson or zero-inflated negative binomial, on a player and overall basis.

Usage

injsummary(
  injd,
  var_type_injury = NULL,
  method = c("poisson", "negbin", "zinfpois", "zinfnb"),
  conf_level = 0.95,
  quiet = FALSE
)

Arguments

injd

injd S3 object (see prepare_all()).

var_type_injury

Character specifying the name of the column according to which compute injury summary statistics. It should refer to a (categorical) variable that describes the "type of injury". Optional, defaults to NULL.

method

Method to estimate injury incidence and injury burden. One of "poisson", "negbin", "zinfpois" or "zinfnb"; characters that stand for Poisson method, negative binomial method, zero-inflated Poisson and zero-inflated negative binomial.

conf_level

Confidence level (default = 0.95).

quiet

Logical, whether or not to silence the warning messages (defaults to FALSE).

Value

A list of two data frames comprising player-wise and overall injury summary statistics, respectively, that constitute an injds

S3 object, and both of them made up of the following columns:

ninjuries: number of injuries sustained by the player over the given period specified by the injd data frame.

ndayslost: number of days lost by the player due to injury over the given period specified by the injd data frame.

totalexpo: total exposure that the player has been under risk of sustaining an injury.

injincidence: injury incidence, number of injuries per unit of exposure.

injburden: injury burden, number of days lost per unit of exposure.

var_type_injury: only if it is specified as an argument to function.

Examples

# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

injsummary(injd)
#> Warning: 
#> * Exposure time unit is matches_minutes
#>   So... Injury incidence and injury burden are calculated per 100 player-matches of exposure (90 minutes times 100)
#> 
#> Warning: 
#> * Exposure time unit is matches_minutes
#>   So... Injury incidence and injury burden are calculated per 100 player-matches of exposure (90 minutes times 100)
#> 
#> $playerwise
#> # A tibble: 28 × 6
#>    player                  ninjuries ndayslost totalexpo injincidence injburden
#>    <fct>                       <dbl>     <dbl>     <dbl>        <dbl>     <dbl>
#>  1 adam-lallana                    6       302       700        77.1     3883. 
#>  2 alberto-moreno                  1        50      1264         7.12     356. 
#>  3 alex-oxlade-chamberlain         1       316      1483         6.07    1918. 
#>  4 alisson                         0         0      3420         0          0  
#>  5 andrew-robertson                5        22      5162         8.72      38.4
#>  6 daniel-sturridge                3       122       927        29.1     1184. 
#>  7 danny-ings                      0         0       265         0          0  
#>  8 dejan-lovren                    6       160      3109        17.4      463. 
#>  9 divock-origi                    1         5       366        24.6      123. 
#> 10 dominic-solanke                 0         0       581         0          0  
#> # … with 18 more rows
#> 
#> $overall
#> # A tibble: 1 × 11
#>   ninjuries ndayslost totalexpo injinc…¹ injbu…² injin…³ injbu…⁴ injin…⁵ injin…⁶
#>       <dbl>     <dbl>     <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1        82      2049       700    1054.  26344.    116.    582.    863.   1246.
#> # … with 2 more variables: injburden_lower <dbl>, injburden_upper <dbl>, and
#> #   abbreviated variable names ¹​injincidence, ²​injburden, ³​injincidence_sd,
#> #   ⁴​injburden_sd, ⁵​injincidence_lower, ⁶​injincidence_upper
#> 
#> attr(,"class")
#> [1] "injds" "list" 
#> attr(,"unit_exposure")
#> [1] "matches_minutes"
#> attr(,"unit_timerisk")
#> [1] "100 player-match"